Trakk ut VelgUstrukturertDokumentasjonType i egen komponent
